+++ lib/nova_ui.dart
@@ -6,6 +6,7 @@ export 'src/widgets/cards/nova_card.dart';
 export 'src/widgets/appbars/nova_app_bar.dart';
 export 'src/widgets/loaders/nova_loader.dart';
 export 'src/widgets/dialogs/nova_dialog.dart';
+export 'src/widgets/badge/nova_badge.dart';
 export 'src/widgets/toasts/nova_toast.dart';
 export 'src/widgets/bottomsheets/nova_bottom_sheet.dart';